requirements.txt
gradio
requests
python-dotenv
newspaper3k
openai